Some space optimizations, and finally using ghost mesh for pacman sample

+++ test/src/me/anno/tests/game/pacman/logic/Node.kt
@@ -1,7 +1,7 @@
 package me.anno.tests.game.pacman.logic
 
-import org.joml.Vector2i
+import org.joml.Vector2f
 
-class Node(val position: Vector2i) {
+class Node(val position: Vector2f) {
     val neighbors = ArrayList<Node>(4)
 }
