+++ test/src/me/anno/tests/game/pacman/logic/Wall.kt
@@ -1,5 +1,5 @@
 package me.anno.tests.game.pacman.logic
 
-import org.joml.Vector2i
+import org.joml.Vector2f
 
-class Wall(val start: Vector2i, val end: Vector2i)
+class Wall(val start: Vector2f, val end: Vector2f)
